directory refactoring
