+++ package/src/scripts/create.ts
