+++ package/src/scripts/destroy.ts
